Improve error handling in TravisStatusHttp
If an error occurs during JSON decoding, we want to return it (but not
in preference to HTTP status errors).

Also stop conflating the Error object with the JSON response body.  We
don't need API compatibility with travis-ci, since we are the only
consumers of the Error object, and this seems much more sane.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>