Restrict keepAlive Agent use
If we are not fetching over HTTP or HTTPS, don't use the http Agent
class.  This is more of a safety check than a real use case, since
request only supports these protocols.  But protocol-relative links or
future support for file:// are conceivable.

Only use the Agent class if it supports keepAlive.  This was introduced
in Node.js 0.11.4.  Before this version, it would needlessly create a
new Agent and stop using the global Agent.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>