Fix and test argument defaulting
Two of the arguments which have default optarg values which should not
be default option values (e.g. HEAD is the default if --commit is given
without an argument, not the default if --commit is not given).  This
code was incorrect and not caught by the test suite.  Fix both.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>