Fix Yaku/Bluebird incompatibility on Node 0.10
On Node 0.10, returning bluebird.reject in the onRejected handler of a
Yaku promise results in unhandledRejection.  I can reproduce the error
with:

    var BBPromise = require('bluebird');
    var YPromise = require('yaku');
    process.on('unhandledRejection', function(err) {
      console.error('unhandledRejection', err);
    });
    YPromise.resolve()
      .then(function() { return BBPromise.reject(new Error()); })
      .catch(function() {});

This prints unhandledRejection on Node 0.10, but not on 0.12 or later.
I think this is due to the process.nextTick callback order, but it's not
worth spending more effort debugging an error that only appears on Node
0.10.  Add an easy workaround.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>