Register bluebird with any-promise when necessary
Tests on Node 0.10 were failing due to native-promise-only being
installed as a transitive dependency, being preferred by any-promise,
and using global setImmediate for its tick (which is overridden by
sinon/lolex for some tests).  Fix it by explicitly registering bluebird
with any-promise in tests when global.Promise is not defined.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>