Improve handling explicit transport in foreign URL
A remote URL may be a "foreign URL" which includes an explicit transport
name followed by "::", as described in
https://git-scm.com/docs/git-remote-helpers

The portion following the URL is passed as an argument to the transport
helper, and may be a path or URL or something else (as with any remote
URL, since transport could also be set by remote.<name>.vcs config).  So
attempt to parse the rest of the URL like any other.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>